# Single pass ReplaceSource .node performance improvement