--- a/lib/ReplaceSource.js
+++ b/lib/ReplaceSource.js
@@ -316,9 +316,6 @@ class ReplacementEnumerator {
 			if(start < 0 && nextReplAtStart) {
 				output.push(repl[2]);
 				this.index++;
-				if(this.index >= this.replacements.length) {
-					this.done = true;
-				}
 			} else {
 				inHeader = false;
 			}