--- a/lib/ReplaceSource.js
+++ b/lib/ReplaceSource.js
@@ -190,8 +190,11 @@ class ReplaceSource extends Source {
 		var splitPosition = replace.position - position;
 		// If multiple replaces occur in the same location then the splitPosition may be
 		// before the current position for the subsequent splits. Ensure it is >= 0
-		if(splitPosition < 0)
+		var originalSplitPosition = 0;
+		if(splitPosition < 0) {
+			originalSplitPosition = splitPosition;
 			splitPosition = 0;
+		}
 		if(splitPosition >= node.length || replace.done) {
 			if(replace.emit) {
 				var nodeEnd = new SourceNode(
@@ -227,10 +230,12 @@ class ReplaceSource extends Source {
 			// then the column position shows the split position of the intermediate source map and not the original
 
 			// Emit the replacement value
-			if(replace.value && replace.value.length > 0) {
+			if(replace.value) {
+				// If the split position was < 0 due to overlapping replaces, adjust the column to match
+				var column = parent.column + originalSplitPosition;
 				output.push(new SourceNode(
 					parent.line,
-					parent.column,
+					column < 0 ? 0 : column,
 					parent.source,
 					replace.value,
 					parent.name
@@ -331,6 +336,8 @@ class ReplacementEnumerator {
 	}
 
 	footer() {
+		if(!this.done && !this.emit)
+			this.next(); // If we finished _replaceInNode mid emit we advance to next entry
 		return this.done ? [] : this.replacements.slice(this.index).map(repl => repl[2]);
 	}
 }