--- a/lib/ReplaceSource.js
+++ b/lib/ReplaceSource.js
@@ -75,7 +75,7 @@ class ReplaceSource extends Source {
 		if(this.replacements.length === 0) {
 			return node;
 		}
-		this._sortReplacementsAscending();
+		this.replacements.sort(sortReplacementsAscending);
 		var replace = new ReplacementEnumerator(this.replacements);
 		var output = [];
 		this._prependStartNodes(output, replace);
@@ -139,18 +139,6 @@ class ReplaceSource extends Source {
 
 	_splitString(str, position) {
 		return position <= 0 ? ["", str] : [str.substr(0, position), str.substr(position)];
-	}
-
-	_sortReplacementsAscending() {
-		this.replacements.sort(function(a, b) {
-			var diff = a[1] - b[1]; // end
-			if(diff !== 0)
-				return diff;
-			diff = a[0] - b[0]; // start
-			if(diff !== 0)
-				return diff;
-			return a[3] - b[3]; // insert order
-		});
 	}
 
 	_replaceInNode(output, replace, node, position) {
@@ -265,6 +253,16 @@ class ReplaceSource extends Source {
 			output.push(remainingValues[i]);
 		}
 	}
+}
+
+function sortReplacementsAscending(a, b) {
+	var diff = a[1] - b[1]; // end
+	if(diff !== 0)
+		return diff;
+	diff = a[0] - b[0]; // start
+	if(diff !== 0)
+		return diff;
+	return a[3] - b[3]; // insert order
 }
 
 class ReplacementEnumerator {