--- a/lib/ReplaceSource.js
+++ b/lib/ReplaceSource.js
@@ -78,10 +78,30 @@ class ReplaceSource extends Source {
 		this.replacements.sort(sortReplacementsAscending);
 		var replace = new ReplacementEnumerator(this.replacements);
 		var output = [];
-		this._prependStartNodes(output, replace);
-		this._replaceInNode(output, replace, node, 0, null);
-		this._appendRemainingNodes(output, replace);
+		var position = 0;
+
+		// We build a new list of SourceNodes in "output"
+		// from the original mapping data
+
+		var replaceInStringNode = this._replaceInStringNode.bind(this, output, replace);
+		node.walk(function(chunk, mapping) {
+			position = replaceInStringNode(chunk, position, mapping);
+		});
+
+		// If any replacements occur after the end of the original file, then we append them
+		// directly to the end of the output
+		var remaining = replace.footer();
+		if(remaining) {
+			output.push(remaining);
+		}
+
 		var result = new SourceNode(null, null, null, output);
+
+		// We need to add source contents afterwards
+		node.walkSourceContents(function(sourceFile, sourceContent) {
+			result.setSourceContent(sourceFile, sourceContent);
+		});
+
 		return result;
 	}
 
@@ -141,117 +161,57 @@ class ReplaceSource extends Source {
 		return position <= 0 ? ["", str] : [str.substr(0, position), str.substr(position)];
 	}
 
-	_replaceInNode(output, replace, node, position) {
-		var outputChildren = [];
-		var allChildrenAreStrings = true;
-
-		for(var i = 0, len = node.children.length; i < len; i++) {
-			var child = node.children[i];
-			if(typeof child !== 'string') {
-				position = this._replaceInNode(outputChildren, replace, child, position);
-				allChildrenAreStrings = false;
-			} else {
-				position = this._replaceInStringNode(outputChildren, replace, child, position, node);
-			}
-		}
-		if(outputChildren.length > 0) {
-			if(allChildrenAreStrings) {
-				for(i = 0; i < outputChildren.length; i++)
-					output.push(outputChildren[i]);
-			} else {
-				var outputNode = new SourceNode(
-					node.line,
-					node.column,
-					node.source,
-					outputChildren,
-					node.name
-				);
-				if(node.sourceContents)
-					outputNode.sourceContents = node.sourceContents;
-				output.push(outputNode);
-			}
-		}
-		return position;
-	}
-
 	_replaceInStringNode(output, replace, node, position, parent) {
-		var splitPosition = replace.position - position;
-		// If multiple replaces occur in the same location then the splitPosition may be
-		// before the current position for the subsequent splits. Ensure it is >= 0
-		var originalSplitPosition = 0;
-		if(splitPosition < 0) {
-			originalSplitPosition = splitPosition;
-			splitPosition = 0;
-		}
-		if(splitPosition >= node.length || replace.done) {
-			if(replace.emit) {
-				var nodeEnd = new SourceNode(
-					parent.line,
-					parent.column,
-					parent.source,
-					node,
-					parent.name
-				);
-				if(parent.sourceContents)
-					nodeEnd.sourceContents = parent.sourceContents;
-				output.push(nodeEnd);
+		do {
+			var splitPosition = replace.position - position;
+			// If multiple replaces occur in the same location then the splitPosition may be
+			// before the current position for the subsequent splits. Ensure it is >= 0
+			if(splitPosition < 0) {
+				splitPosition = 0;
 			}
-			return(position + node.length);
-		}
-		var emit = replace.next();
-		if(!emit) {
-			// Stop emitting when we have found the beginning of the string to replace.
-			// Emit the part of the string before splitPosition
-			var nodeStart = new SourceNode(
-				parent.line,
-				parent.column,
-				parent.source,
-				node.substr(0, splitPosition),
-				parent.name
-			);
-			if(parent.sourceContents)
-				nodeStart.sourceContents = parent.sourceContents;
-			output.push(nodeStart);
-
-			// We should advance the current column position by splitPosition characters at this point.
-			// However if multiple ReplaceSource's are chained (as occurs when using ModuleConcatenationPlugin)
-			// then the column position shows the split position of the intermediate source map and not the original
-
-			// Emit the replacement value
-			if(replace.value) {
-				// If the split position was < 0 due to overlapping replaces, adjust the column to match
-				var column = parent.column + originalSplitPosition;
-				output.push(new SourceNode(
-					parent.line,
-					column < 0 ? 0 : column,
-					parent.source,
-					replace.value,
-					parent.name
-				));
+			if(splitPosition >= node.length || replace.done) {
+				if(replace.emit) {
+					var nodeEnd = new SourceNode(
+						parent.line,
+						parent.column,
+						parent.source,
+						node,
+						parent.name
+					);
+					output.push(nodeEnd);
+				}
+				return position + node.length;
 			}
-		}
+			var emit = replace.next();
+			if(!emit) {
+				// Stop emitting when we have found the beginning of the string to replace.
+				// Emit the part of the string before splitPosition
+				if(splitPosition > 0) {
+					var nodeStart = new SourceNode(
+						parent.line,
+						parent.column,
+						parent.source,
+						node.substr(0, splitPosition),
+						parent.name
+					);
+					output.push(nodeStart);
+				}
 
-		// Recurse with remainder of the string as there may be multiple replaces within a single node
-		var remainder = node.substr(splitPosition, node.length);
-		return this._replaceInStringNode(output, replace, remainder, position + splitPosition, parent);
-	}
+				// Emit the replacement value
+				if(replace.value) {
+					output.push(new SourceNode(
+						parent.line,
+						parent.column,
+						parent.source,
+						replace.value
+					));
+				}
+			}
 
-	_prependStartNodes(output, replace) {
-		// If any replacements occur before the start of the original file, then we prepend them
-		// directly to the start of the output
-		var startValues = replace.header();
-		for(var i = 0; i < startValues.length; i++) {
-			output.push(startValues[i]);
-		}
-	}
-
-	_appendRemainingNodes(output, replace) {
-		// If any replacements occur after the end of the original file, then we append them
-		// directly to the end of the output
-		var remainingValues = replace.footer();
-		for(var i = 0; i < remainingValues.length; i++) {
-			output.push(remainingValues[i]);
-		}
+			// Recurse with remainder of the string as there may be multiple replaces within a single node
+			node = node.substr(splitPosition);
+			position += splitPosition;
+		} while (true);
 	}
 }
 
@@ -306,37 +266,12 @@ class ReplacementEnumerator {
 		return this.emit;
 	}
 
-	header() {
-		var output = [];
-		var inHeader = true;
-		while(inHeader && !this.done) {
-			var repl = this.replacements[this.index];
-			var start = Math.floor(repl[0]);
-			this.position = start;
-
-			// Replicate previous ReplaceSource behavior:
-			// In order to generate identical source maps to the previous version of webpack we should not output the last header
-			// node as a string if the next replace starts above position 0
-			var nextReplAtStart = this.index == (this.replacements.length - 1) ?
-				false :
-				Math.floor(this.replacements[this.index + 1][0]) <= 0;
-
-			if(start < 0 && nextReplAtStart) {
-				output.push(repl[2]);
-				this.index++;
-			} else {
-				inHeader = false;
-			}
-		}
-		if(this.position < 0)
-			this.position = 0;
-		return output;
-	}
-
 	footer() {
 		if(!this.done && !this.emit)
 			this.next(); // If we finished _replaceInNode mid emit we advance to next entry
-		return this.done ? [] : this.replacements.slice(this.index).map(repl => repl[2]);
+		return this.done ? [] : this.replacements.slice(this.index).map(function(repl) {
+			return repl[2];
+		}).join("");
 	}
 }