**What kind of change does this PR introduce?**
Performance improvement. Up to 35% faster webpack builds on large solutions.
(small solutions do not spend as much time in ReplaceSource)

**Did you add tests for your changes?**
Added tests to improve coverage around prepending and appending replaces as these were not covered by existing tests.

**If relevant, link to documentation update:**
N/A

**Summary**
Replaces the .node method in ReplaceSource with a version which performs all replacements in a single pass.

**Does this PR introduce a breaking change?**
No

**Other information**
This PR can produce significantly faster build times when processing large files with devtool: 'source-map'

The existing .node method walks the SourceNode tree for every replace each time, twice splitting the SourceNode tree at the replace start and end. In large source maps this can take significant time as the SourceNode constructor will individually .add each child node on each split.

The new method processes all replaces inline, turning on and off an emit flag depending on if you are in the middle of a replace or not.
The PR emits identical module and source map output as webpack 3.4.1 / webpack-sources 1.0.1. (Including use with ModuleConcatenationPlugin which chains ReplaceSource's)
 
For our current Angular prod build using devtool: 'source-map' and 2481 modules, this PR reduces the _chunk asset optimization_ time (average over 3 runs) from ~314 seconds to ~127 seconds, and reduces total webpack build time from ~485 seconds to ~314 seconds. (38.5% or 187 seconds faster)