@brentwilton 2017-08-14
@sokra @TheLarkInn please let me know if there is anything additional I can provide to help clarify the issue or to explain how the updated ReplaceSource.node() method works.

We are keen to get this PR merged as it makes our CI webpack builds 3 minutes faster.