@brentwilton 2017-11-30
@alan-agius4 We have been using a postinstall script in the package.json to manually patch ReplaceSource.js for the last few months while we have been waiting for this PR to be merged.
(I highly recommend implementing this workaround if you have not already. The performance benefits outweigh the hackiness)

With webpack 3.8.1 and the 1.1.0 version of ReplaceSource our CI build takes **47 minutes**. With the PR version, it takes **17 minutes**. (the patch version was temporarily not being applied on our CI build yesterday and everyone thought that the CI server had broken due to builds taking forever)