@alan-agius4 2017-11-30
I dont understand and ita frustrating why this PR is not being prioritized! I have the same problem. My CI build iktakes approx 40mins because of this method.