@sokra 2017-12-01
This is difficult to review, but I take a look soon.