@brentwilton 2017-12-02
Thanks for the update @sokra. I know you guys have been busy the last few months with progressing webpack 4.

To give a bit more information about the cause of the slowdown and the solution:

The current algorithm does the following:
For _each_ replace we:
1. Traverse the entire source map tree until we find the replace end point
2. Split the source map tree at that point
3. Traverse the entire source map tree until we find the replace start point
4. Split the source map tree at that point
5. Add the new replace text and the end section
6. Move onto the next replacement with the remainder of the tree
		
There are some performance issues with this approach
- We traverse the entire remaining tree twice per replace
- The tree is split twice per replace, at the start and end point of the replace
- When the tree is split, it adds the split sections as children to 2 new SourceNodes
- When you create a new SourceNode, internally it will type check and array.push each child individually to its children array
- On large source maps with a lot of replaces this can take minutes and it thrashes memory while doing so.

The new algorithm changes this to only traverse the source node tree a single time making a copy of it to a new tree.

1. We first sort the list of replacements by start location, and find the first location where we want to replace.
2. When traversing the tree we look for the first replacements start point, and when we hit that point, we turn an emit flag off and stop copying across to the new tree.
3. We add the replacement text as the next node in the copy.
4. We then continue the traversal and look for the replacement end point. On hitting the end point we turn back	on the emit flag and start copying the tree again.
5. When we hit the end point and start emitting again, we then move onto the next replacement in the list and repeat the above until the entire tree is copied.

The ReplacementEnumerator class handles the state of the emit flag and current replacement index.
	
There are some special cases we also deal with:
- Some of the replacements occur outside of the original source map. This is usually webpack adding the imports or require wrappers. We run the _prependStartNodes and _appendRemainingNodes steps to handle this and make the main traversal copy code less complex.
- Replaces can only occur on leaf nodes, and multiple replaces can occur on the same leaf node. Sometimes the start/end points will even overlap so we actually keep recursing into replaceInStringNode after we finish a replace to handle this.
- If you look in the header() method I included some code which tweaks the output to generate identical source maps to the previous version. Identical source map output is good to show we didn't break anything in the change.

We had a situation where a single ReplaceSource.node call was taking 15 minutes on a complex build. The new algorithm reduced that call to 0.12 seconds.