@vladimir-tikhonov 2017-12-08
Hello, @brentwilton! I've tried use your branch on our project, but unfortunately found a bug. We use `redux-action` package and now webpack produces invalid js code when `devtool` is set to `"source-map"`. Seems like this file is causing issues: https://github.com/reduxactions/redux-actions/blob/master/src/camelCase.js. I've created a repro repository for you here: https://github.com/vladimir-tikhonov/webpack-sources-bug. Let me know if you need something else.