@brentwilton 2017-12-10
Hi @vladimir-tikhonov. Thanks for trying this PR out and submitting a repro of your issue.
I will pull it down and isolate the cause.