@brentwilton 2017-12-11
@vladimir-tikhonov  I have pushed a new version. Can you check to see if it resolves your issue ?

The code was double emitting the `);` if it reached the end of the source map while in the middle of a replace.