@vladimir-tikhonov 2017-12-11
@brentwilton yes, the issue is resolved now. Great job!