@TheLarkInn 2018-01-10
Yes cc @sokra and @ooflorent also.

On Sun, Jan 7, 2018, 12:30 PM Alan Agius <notifications@github.com> wrote:

> Can we have this re-reviewed please?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/webpack/webpack-sources/pull/23#issuecomment-355850264>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADQBMAoKsgJ51a4d4QVMv4ra-hFecZjDks5tISl3gaJpZM4Oe2IV>
> .
>