@TheLarkInn 2018-01-25
I think we are waiting on @ooflorent's request, and I'd like @sokra to review as well for this. Given we just released webpack@4.0.0-beta.0 we need to make sure this doesn't cause any breaking changes (even if tests passing).