@grahamlucid 2018-04-12
Is there any update on getting this issue merged?

Thanks!