@rpellerin 2018-05-02
Would love to see this merged too!