@mlavina 2018-08-27
Hey @sokra @ooflorent 

Just checking in as well if this can be merged in and released soon?