@sokra 2018-08-29
and released as `1.2.0`