@alan-agius4 2018-08-29
thank you :)

On Wed, 29 Aug 2018 at 14:10, Tobias Koppers <notifications@github.com>
wrote:

> and released as 1.2.0
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/webpack/webpack-sources/pull/23#issuecomment-416929602>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AQv-WteNxcCLROvWIsoYIWXlsaC3wC99ks5uVoShgaJpZM4Oe2IV>
> .
>