@eranimo 2018-08-29
Is it possible to create a beta release version of webpack with this new version?