@edmorley 2018-08-29
This decreased the time taken for our warm (babel-loader and uglify-webpack-plugin caches enabled) webpack 4 `--mode production` `devtool: 'source-map'` builds by 30%. Much appreciated - thank you :-)